+++ Changelog.md
@@ -1,1 +1,1 @@
-- Fixed turbine tooltip
+- Updated melter recipes